fix extra pedantic warnings on gcc-7...gcc-10

+++ src/c4/memory_resource.hpp
@@ -386,12 +386,12 @@ struct AllocationCounts
         void add(size_t sz)
         {
             ++allocs;
-            size += sz;
+            size += static_cast<ssize_t>(sz);
         }
         void rem(size_t sz)
         {
             --allocs;
-            size -= sz;
+            size -= static_cast<ssize_t>(sz);
         }
         Item max(Item const& that) const
         {
